Flush Stdin on error
Without this change, when the user enters a large number of characters
into the string field for scanf a load of garbage is printed to the
console as stdin still retains all the extraneous characters inputted
earlier. Specifically the user experiences many prompts entered with
chunks of those large strings (entered as input to the prompt)
cluttering the screen and confusing the user.